refactor: extract line classification and grouping to core lexing modules
Extracts core lexer logic from transformation layer into dedicated modules:
- line_classification.rs: Line type classification logic and helpers
- line_grouping.rs: Token grouping logic that calls classifiers

The transformation now only handles stream mapping, delegating to core
lexer modules. This improves separation of concerns and makes the core
classification logic more accessible.

Changes:
- Add src/lex/lexing/line_classification.rs (314 lines)
- Add src/lex/lexing/line_grouping.rs (260 lines)
- Refactor transformations/line_token_grouping.rs (629→83 lines)
- Update module declarations in lexing.rs

Net addition: 28 lines (mostly module headers and imports)
All tests passing: 406 tests, 0 failures
